+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 # 1.0.0+ - 2019-09-23
 ## Added
+- 'https://' is now optional for `--server`
 
 ## Fixed
 - `enable = false` in `git config` is no longer treated as a True value